Moved common functions into liscan.py Adapted the modules and tests accordingly Added one atom to the pyrimidine to purine mutations to use three common atoms (after renaming) for the base reconstruction
